Fix Android native bridges for Conversations 6.2.1 and Pigeon 29 events.
Align Kotlin mappers and emitters with Twilio SDK 6.2.1 so CI unit tests compile and pass, including JVM-safe attribute JSON validation.

+++ packages/twilio_flutter_conversations/android/src/main/kotlin/com/twilioflutter/conversations/bridge/MessageMediaMapping.kt
@@ -13,7 +13,7 @@ object MessageMediaMapping {
       sid = media.sid,
       contentType = media.contentType ?: "application/octet-stream",
       filename = media.filename ?: "",
-      sizeBytes = media.size.toInt(),
+      sizeBytes = media.size,
     )
   }
